some more comments restructured teh package info
