+++ src/main/java/edu/kit/trufflehog/command/usercommand/SetDisplayAlgorithmCommand.java
@@ -1,7 +1,7 @@
 package edu.kit.trufflehog.command.usercommand;
 
-import edu.kit.trufflehog.model.graph.IConnection;
-import edu.kit.trufflehog.model.graph.INode;
+import edu.kit.trufflehog.model.network.graph.IConnection;
+import edu.kit.trufflehog.model.network.graph.INode;
 import edu.kit.trufflehog.model.network.INetworkViewPort;
 import edu.uci.ics.jung.algorithms.layout.Layout;
 import edu.uci.ics.jung.graph.Graph;
