+++ src/main/java/edu/kit/trufflehog/model/network/graph/FRLayoutFactory.java
@@ -1,7 +1,6 @@
-package edu.kit.trufflehog.model.graph;
+package edu.kit.trufflehog.model.network.graph;
 
-import edu.kit.trufflehog.model.graph.jungconcurrent.ConcurrentFRLayout;
-import edu.uci.ics.jung.algorithms.layout.FRLayout;
+import edu.kit.trufflehog.model.network.graph.jungconcurrent.ConcurrentFRLayout;
 import edu.uci.ics.jung.algorithms.layout.Layout;
 import edu.uci.ics.jung.graph.Graph;
 import org.apache.commons.collections15.Transformer;
