+++ src/main/java/edu/kit/trufflehog/model/network/graph/jungconcurrent/ConcurrentAbstractLayout.java
@@ -1,4 +1,4 @@
-package edu.kit.trufflehog.model.graph.jungconcurrent;
+package edu.kit.trufflehog.model.network.graph.jungconcurrent;
 
 import edu.uci.ics.jung.algorithms.layout.Layout;
 import edu.uci.ics.jung.graph.Graph;
