+++ src/main/java/edu/kit/trufflehog/model/network/recording/INetworkReadingPortSwitch.java
@@ -1,7 +1,14 @@
 package edu.kit.trufflehog.model.network.recording;
 
+import edu.kit.trufflehog.model.network.INetworkReadingPort;
+
 /**
  * Created by jan on 20.02.16.
  */
-public interface INetworkReadingPortSwitch {
+public interface INetworkReadingPortSwitch extends INetworkReadingPort {
+
+    INetworkReadingPort getActiveReadingPort();
+
+    void setActiveReadingPort(INetworkReadingPort port);
+
 }
