+++ src/main/java/edu/kit/trufflehog/model/network/recording/INetworkWritingPortSwitch.java
@@ -1,7 +1,14 @@
 package edu.kit.trufflehog.model.network.recording;
 
+import edu.kit.trufflehog.model.network.INetworkWritingPort;
+
 /**
- * Created by jan on 20.02.16.
+ *
  */
-public interface INetworkWritingPortSwitch {
+public interface INetworkWritingPortSwitch extends INetworkWritingPort {
+
+    INetworkWritingPort getActiveWritingPort();
+
+    void setActiveWritingPort(INetworkWritingPort writingPort);
+
 }
