+++ src/main/java/edu/kit/trufflehog/model/network/recording/NetworkReadingPortSwitch.java
@@ -3,7 +3,7 @@ package edu.kit.trufflehog.model.network.recording;
 import edu.kit.trufflehog.model.network.INetworkReadingPort;
 
 /**
- * Created by jan on 20.02.16.
+ * TODO implement
  */
 public class NetworkReadingPortSwitch implements INetworkReadingPortSwitch {
     public NetworkReadingPortSwitch(INetworkReadingPort readingPort) {
