+++ src/main/java/edu/kit/trufflehog/model/network/recording/NetworkWritingPortSwitch.java
@@ -2,9 +2,7 @@ package edu.kit.trufflehog.model.network.recording;
 
 import edu.kit.trufflehog.model.network.INetworkWritingPort;
 
-/**
- * Created by jan on 20.02.16.
- */
+// TODO implement
 public class NetworkWritingPortSwitch implements INetworkWritingPortSwitch {
     public NetworkWritingPortSwitch(INetworkWritingPort writingPort) {
     }
