+++ src/main/java/edu/kit/trufflehog/model/network/recording/ReplayNetwork.java
@@ -1,7 +1,7 @@
 package edu.kit.trufflehog.model.network.recording;
 
-import edu.kit.trufflehog.model.graph.IConnection;
-import edu.kit.trufflehog.model.graph.INode;
+import edu.kit.trufflehog.model.network.graph.IConnection;
+import edu.kit.trufflehog.model.network.graph.INode;
 import edu.kit.trufflehog.model.network.INetwork;
 import edu.kit.trufflehog.model.network.INetworkReadingPort;
 import edu.kit.trufflehog.model.network.INetworkViewPort;
@@ -9,7 +9,8 @@ import edu.kit.trufflehog.model.network.INetworkWritingPort;
 import edu.uci.ics.jung.graph.Graph;
 
 /**
- * Created by jan on 20.02.16.
+ * TODO IMPLEMENT having replay network is necessary, because there are several internal functionalities that
+ * TODO are handled a bite differently in contrast to a live network
  */
 public class ReplayNetwork implements INetwork {
 
@@ -18,6 +19,7 @@ public class ReplayNetwork implements INetwork {
 
     @Override
     public INetworkReadingPort getReadingPort() {
+
         return null;
     }
