+++ src/main/java/edu/kit/trufflehog/view/NetworkGraphView.java
@@ -2,8 +2,8 @@ package edu.kit.trufflehog.view;
 
 import edu.kit.trufflehog.command.usercommand.IUserCommand;
 import edu.kit.trufflehog.interaction.GraphInteraction;
-import edu.kit.trufflehog.model.graph.IConnection;
-import edu.kit.trufflehog.model.graph.INode;
+import edu.kit.trufflehog.model.network.graph.IConnection;
+import edu.kit.trufflehog.model.network.graph.INode;
 import edu.kit.trufflehog.model.network.INetworkViewPort;
 import edu.kit.trufflehog.view.controllers.NetworkGraphViewController;
 import edu.uci.ics.jung.algorithms.layout.GraphElementAccessor;
