+++ src/main/java/edu/kit/trufflehog/view/controllers/NetworkGraphViewController.java
@@ -2,8 +2,8 @@ package edu.kit.trufflehog.view.controllers;
 
 import edu.kit.trufflehog.command.usercommand.IUserCommand;
 import edu.kit.trufflehog.interaction.GraphInteraction;
-import edu.kit.trufflehog.model.graph.IConnection;
-import edu.kit.trufflehog.model.graph.INode;
+import edu.kit.trufflehog.model.network.graph.IConnection;
+import edu.kit.trufflehog.model.network.graph.INode;
 import edu.kit.trufflehog.util.IListener;
 import edu.kit.trufflehog.util.INotifier;
 import edu.uci.ics.jung.visualization.VisualizationServer;
